tidy up missing parts of new shared library build

+++ builds/build_android_abis.sh
@@ -5,7 +5,9 @@ set -e
 ABIS="x86 x86_64 armeabi-v7a arm64-v8a"
 for abi in $ABIS
 do
-    rm -r ./android
+    if [ -d "./android" ]; then
+        rm -r ./android
+    fi
     export ANDROID_ABI=$abi
     ./cmake_lib_android.sh
     (cd android/; make)
